This fixes an issue with logout now bringing back to login screen.
tested on POP_OS

Issue: pressing the logout button on polybar does no bring back to login screen. It instead brings the user to the BIOS screen and stays frozen there.

Fix: By doing this change I was able to fix the issue. Logout now sends the user back to the login screen.